backend/requirements.txt
fastapi==0.99.1
uvicorn==0.22.0
PyJWT==2.7.0
python-decouple==3.8
motor == 3.2.0
certifi == 2023.5.7
email-validator == 2.0.0.post2
Werkzeug == 2.3.6